Changed all size_t to uint64_t due to overflow

+++ core/inc/artsGpuStream.h
@@ -71,8 +71,8 @@ typedef struct
 typedef struct 
 {
     int device;
-    volatile size_t availGlobalMem;
-    volatile size_t totalGlobalMem;
+    volatile uint64_t availGlobalMem;
+    volatile uint64_t totalGlobalMem;
     struct cudaDeviceProp prop;
     volatile float occupancy;
     volatile unsigned int deviceLock;
@@ -109,7 +109,7 @@ extern artsGpu_t * artsGpus;
 
 extern volatile unsigned int hits;
 extern volatile unsigned int misses;
-extern volatile unsigned int freeBytes;
+extern volatile uint64_t freeBytes;
 
 #ifdef __cplusplus
 }
